# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "Hostaway Phonebook API. Test API token for authorization '24f56647eddc650bd0904883dd7168e609017696cf69714fe7d1224012491710'.",
        "contact": {
            "email": "george.webfullstack@gmail.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Api Server"
        }
    ],
    "paths": {
        "/api/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get items",
                "description": "Returns items from phonebook all or by query",
                "operationId": "App\\Http\\Controllers\\ItemController::index",
                "parameters": [
                    {
                        "name": "q",
                        "in": "path",
                        "description": "Query text",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Item get error"
                    }
                },
                "security": [
                    {
                        "apiToken": [
                            "read:items"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Create item",
                "description": "Create item in phonebook",
                "operationId": "App\\Http\\Controllers\\ItemController::store",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "path",
                        "description": "Item first name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "path",
                        "description": "Item last name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "path",
                        "description": "Item uniq phone number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country_code",
                        "in": "path",
                        "description": "Item country code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "timezone_name",
                        "in": "path",
                        "description": "Item timezone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Item create error"
                    }
                },
                "security": [
                    {
                        "apiToken": [
                            "write:items"
                        ]
                    }
                ]
            }
        },
        "/api/items/{id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get item by id",
                "description": "Returns item from phonebook",
                "operationId": "App\\Http\\Controllers\\ItemController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Item id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "422": {
                        "description": "Item get error"
                    }
                },
                "security": [
                    {
                        "apiToken": [
                            "read:items"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Delete item by id",
                "description": "Delete item from phonebook",
                "operationId": "App\\Http\\Controllers\\ItemController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Item id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "422": {
                        "description": "Item delete error"
                    }
                },
                "security": [
                    {
                        "apiToken": [
                            "read:items"
                        ]
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Items"
                ],
                "summary": "Update item by id",
                "description": "Update item in phonebook",
                "operationId": "App\\Http\\Controllers\\ItemController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Item id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "path",
                        "description": "Item first name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "path",
                        "description": "Item last name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "path",
                        "description": "Item uniq phone number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country_code",
                        "in": "path",
                        "description": "Item country code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "timezone_name",
                        "in": "path",
                        "description": "Item timezone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Item not found"
                    },
                    "422": {
                        "description": "Item update error"
                    }
                },
                "security": [
                    {
                        "apiToken": [
                            "read:items"
                        ]
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiToken": {
                "type": "apiKey",
                "name": "api_token",
                "in": "query"
            }
        }
    },
    "tags": [
        {
            "name": "Items",
            "description": "Hostaway phonebook items"
        }
    ],
    "security": [
        []
    ]
}